requirements.txt
requests
automata-lib==1.0.0.post4
nltk